move pck_acc
