+++ src/otx/algo/keypoint_detection/utils/keypoint_eval.py
@@ -8,6 +8,7 @@ from itertools import product
 
 import cv2
 import numpy as np
+from otx.core.metrics.pck import keypoint_pck_accuracy
 
 
 def get_simcc_maximum(
@@ -76,200 +77,6 @@ def get_simcc_maximum(
         vals = vals.reshape(batch_size, num_keypoints)
 
     return locs, vals
-
-
-def get_heatmap_maximum(heatmaps: np.ndarray) -> tuple[np.ndarray, np.ndarray]:
-    """Get maximum response location and value from heatmaps.
-
-    Note:
-        batch_size: B
-        num_keypoints: K
-        heatmap height: H
-        heatmap width: W
-
-    Args:
-        heatmaps (np.ndarray): Heatmaps in shape (K, H, W) or (B, K, H, W)
-
-    Returns:
-        tuple:
-        - locs (np.ndarray): locations of maximum heatmap responses in shape
-            (K, 2) or (B, K, 2)
-        - vals (np.ndarray): values of maximum heatmap responses in shape
-            (K,) or (B, K)
-    """
-    if heatmaps.ndim != 3 or heatmaps.ndim != 4:
-        msg = f"Invalid shape {heatmaps.shape}"
-        raise ValueError(msg)
-
-    if heatmaps.ndim == 3:
-        num_keypoints, y_dim, x_dim = heatmaps.shape
-        batch_size = None
-        heatmaps_flatten = heatmaps.reshape(num_keypoints, -1)
-    else:
-        batch_size, num_keypoints, y_dim, x_dim = heatmaps.shape
-        heatmaps_flatten = heatmaps.reshape(batch_size * num_keypoints, -1)
-
-    y_locs, x_locs = np.unravel_index(np.argmax(heatmaps_flatten, axis=1), shape=(y_dim, x_dim))
-    locs = np.stack((x_locs, y_locs), axis=-1).astype(np.float32)
-    vals = np.amax(heatmaps_flatten, axis=1)
-    locs[vals <= 0.0] = -1
-
-    if batch_size:
-        locs = locs.reshape(batch_size, num_keypoints, 2)
-        vals = vals.reshape(batch_size, num_keypoints)
-
-    return locs, vals
-
-
-def _calc_distances(preds: np.ndarray, gts: np.ndarray, mask: np.ndarray, norm_factor: np.ndarray) -> np.ndarray:
-    """Calculate the normalized distances between preds and target.
-
-    Note:
-        - instance number: N
-        - keypoint number: K
-        - keypoint dimension: D (normally, D=2 or D=3)
-
-    Args:
-        preds (np.ndarray[N, K, D]): Predicted keypoint location.
-        gts (np.ndarray[N, K, D]): Groundtruth keypoint location.
-        mask (np.ndarray[N, K]): Visibility of the target. False for invisible
-            joints, and True for visible. Invisible joints will be ignored for
-            accuracy calculation.
-        norm_factor (np.ndarray[N, D]): Normalization factor.
-            Typical value is heatmap_size.
-
-    Returns:
-        np.ndarray[K, N]: The normalized distances. \
-            If target keypoints are missing, the distance is -1.
-    """
-    batch_size, num_keypoints, _ = preds.shape
-    # set mask=0 when norm_factor==0
-    _mask = mask.copy()
-    _mask[np.where((norm_factor == 0).sum(1))[0], :] = False
-
-    distances = np.full((batch_size, num_keypoints), -1, dtype=np.float32)
-    # handle invalid values
-    norm_factor[np.where(norm_factor <= 0)] = 1e6
-    distances[_mask] = np.linalg.norm(((preds - gts) / norm_factor[:, None, :])[_mask], axis=-1)
-    return distances.T
-
-
-def _distance_acc(distances: np.ndarray, thr: float = 0.5) -> float | int:
-    """Return the percentage below the distance threshold.
-
-    Ignore distances values with -1.
-
-    Note:
-        - instance number: N
-
-    Args:
-        distances (np.ndarray[N, ]): The normalized distances.
-        thr (float): Threshold of the distances.
-
-    Returns:
-        float: Percentage of distances below the threshold. \
-            If all target keypoints are missing, return -1.
-    """
-    distance_valid = distances != -1
-    num_distance_valid = distance_valid.sum()
-    if num_distance_valid > 0:
-        return (distances[distance_valid] < thr).sum() / num_distance_valid
-    return -1
-
-
-def keypoint_pck_accuracy(
-    pred: np.ndarray,
-    gt: np.ndarray,
-    mask: np.ndarray,
-    thr: np.ndarray,
-    norm_factor: np.ndarray,
-) -> tuple[np.ndarray, float, int]:
-    """Calculate the pose accuracy of PCK for each individual keypoint.
-
-    And the averaged accuracy across all keypoints for coordinates.
-
-    Note:
-        PCK metric measures accuracy of the localization of the body joints.
-        The distances between predicted positions and the ground-truth ones
-        are typically normalized by the bounding box size.
-        The threshold (thr) of the normalized distance is commonly set
-        as 0.05, 0.1 or 0.2 etc.
-
-        - instance number: N
-        - keypoint number: K
-
-    Args:
-        pred (np.ndarray[N, K, 2]): Predicted keypoint location.
-        gt (np.ndarray[N, K, 2]): Groundtruth keypoint location.
-        mask (np.ndarray[N, K]): Visibility of the target. False for invisible
-            joints, and True for visible. Invisible joints will be ignored for
-            accuracy calculation.
-        thr (float): Threshold of PCK calculation.
-        norm_factor (np.ndarray[N, 2]): Normalization factor for H&W.
-
-    Returns:
-        tuple: A tuple containing keypoint accuracy.
-
-        - acc (np.ndarray[K]): Accuracy of each keypoint.
-        - avg_acc (float): Averaged accuracy across all keypoints.
-        - cnt (int): Number of valid keypoints.
-    """
-    distances = _calc_distances(pred, gt, mask, norm_factor)
-    acc: np.ndarray = np.array([_distance_acc(d, thr) for d in distances])
-    valid_acc: np.ndarray = acc[acc >= 0]
-    cnt: int = len(valid_acc)
-    avg_acc: float = valid_acc.mean() if cnt > 0 else 0.0
-    return acc, avg_acc, cnt
-
-
-def pose_pck_accuracy(
-    output: np.ndarray,
-    target: np.ndarray,
-    mask: np.ndarray,
-    thr: float = 0.05,
-    normalize: np.ndarray | None = None,
-) -> tuple:
-    """Calculate the pose accuracy of PCK for each individual keypoint.
-
-    And the averaged accuracy across all keypoints from heatmaps.
-
-    Note:
-        PCK metric measures accuracy of the localization of the body joints.
-        The distances between predicted positions and the ground-truth ones
-        are typically normalized by the bounding box size.
-        The threshold (thr) of the normalized distance is commonly set
-        as 0.05, 0.1 or 0.2 etc.
-
-        - batch_size: N
-        - num_keypoints: K
-        - heatmap height: H
-        - heatmap width: W
-
-    Args:
-        output (np.ndarray[N, K, H, W]): Model output heatmaps.
-        target (np.ndarray[N, K, H, W]): Groundtruth heatmaps.
-        mask (np.ndarray[N, K]): Visibility of the target. False for invisible
-            joints, and True for visible. Invisible joints will be ignored for
-            accuracy calculation.
-        thr (float): Threshold of PCK calculation. Default 0.05.
-        normalize (np.ndarray[N, 2]): Normalization factor for H&W.
-
-    Returns:
-        tuple: A tuple containing keypoint accuracy.
-
-        - np.ndarray[K]: Accuracy of each keypoint.
-        - float: Averaged accuracy across all keypoints.
-        - int: Number of valid keypoints.
-    """
-    batch_size, num_keypoints, y_dim, x_dim = output.shape
-    if num_keypoints == 0:
-        return None, 0, 0
-    if normalize is None:
-        normalize = np.tile(np.array([[y_dim, x_dim]]), (batch_size, 1))
-
-    pred, _ = get_heatmap_maximum(output)
-    gt, _ = get_heatmap_maximum(target)
-    return keypoint_pck_accuracy(pred, gt, mask, thr, normalize)
 
 
 def simcc_pck_accuracy(
